requirements.txt
requests
sklearn-pandas==2.2.0
tensorflow
keras
opencv-python
flask
Pillow
